src/requirements.txt
streamlit
langchain
langchain-openai
langgraph
langchain-community
yfinance
chromadb
pypdf
tiktoken
python-dotenv
